normalise node's profit margin during rewarding
